requirements.txt
Django==1.10.3
django-crispy-forms==1.6.1
django-mathfilters==0.4.0
flake8==3.0.4
isort==4.2.5
